requirements.txt
requests==2.31.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
python-dotenv==1.0.0
plotly==5.18.0
dash==2.14.2
folium==0.15.1
altair==5.2.0 